+++ Segundo Bimestre/ProyectoNegocios/src/app/rutas/ruta-aprobar-negocios/ruta-aprobar-negocios.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import {NegocioAPIService} from "../../servicios/api/negocio/negocio-api.service";
 import {AuthService} from "../../servicios/autenticacion/autenticacion.service";
 import {NegocioModelo} from "../../modelos/negocio.modelo";
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-ruta-aprobar-negocios',
@@ -32,19 +33,54 @@ export class RutaAprobarNegociosComponent implements OnInit {
   }
 
   aprobarNegocio(negocio: NegocioModelo) {
-    this.negocioAPIService.updateEstadoNegocio(negocio.id_negocio, this.ESTADO_ACTIVO)
-      .then(resultadoActualizacion => {
-        // TODO: Alerta => se ha activado/actualizado el negocio
-        this.actualizarNegociosPendientes();
-      })
+    Swal.fire({
+      title: '¿Desea activar el negocio '+ negocio.nombre +'?',
+      text: "Confirmar acción",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Activar',
+      cancelButtonText: 'No',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.negocioAPIService.updateEstadoNegocio(negocio.id_negocio, this.ESTADO_ACTIVO)
+          .then(resultadoActualizacion => {
+            Swal.fire(
+              'Activado',
+              '',
+              'success'
+            )
+            this.actualizarNegociosPendientes();
+          })
+      }
+    });
   }
 
   rechazarNegocio(negocio: NegocioModelo) {
-    // TODO: Alerta => Esta seguro?
-    this.negocioAPIService.deleteNegocioPorID(negocio.id_negocio)
-      .then(resultadoEliminacion => {
-        this.actualizarNegociosPendientes();
-      })
+    Swal.fire({
+      title: '¿Está seguro de querer rechazar el negocio '+ negocio.nombre +'?',
+      text: "Esta acción no se puede deshacer",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Rechazar',
+      cancelButtonText: 'No',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.negocioAPIService.deleteNegocioPorID(negocio.id_negocio)
+          .then(resultadoEliminacion => {
+            Swal.fire(
+              'Rechazado!',
+              '',
+              'success'
+            )
+            this.actualizarNegociosPendientes();
+          })
+      }
+    })
+
   }
 
 }
