+++ Segundo Bimestre/ProyectoNegocios/src/app/rutas/ruta-negocios-propios/ruta-negocios-propios.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import {NegocioAPIService} from "../../servicios/api/negocio/negocio-api.service";
 import {NegocioModelo} from "../../modelos/negocio.modelo";
 import {AuthService} from "../../servicios/autenticacion/autenticacion.service";
+import Swal from "sweetalert2";
 
 @Component({
   selector: 'app-ruta-negocios-propios',
@@ -31,27 +32,57 @@ export class RutaNegociosPropiosComponent implements OnInit {
   }
 
   cambiarEstadoNegocio(negocio: NegocioModelo) {
-    // TODO: Alerta => Quiere activar/inactivar el negocio 'NOMBRE' ??/
     const nuevoEstado = negocio.estado == this.ESTADO_ACTIVO ? this.ESTADO_INACTIVO : this.ESTADO_ACTIVO;
-    this.negocioAPIService.updateEstadoNegocio(negocio.id_negocio, nuevoEstado)
-      .then(resultadoActualizacion => {
-        console.error(resultadoActualizacion.error);
-      });
-    negocio.estado = nuevoEstado;
+    Swal.fire({
+      title: '¿Está seguro de querer cambiar el estado a'+ nuevoEstado +'?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí',
+      cancelButtonText: 'No',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.negocioAPIService.updateEstadoNegocio(negocio.id_negocio, nuevoEstado)
+          .then(resultadoActualizacion => {
+            negocio.estado = nuevoEstado;
+            Swal.fire(
+              'Actualizado!',
+              'Cambiado a ' + nuevoEstado,
+              'success'
+            )
+          });
+      }
+    })
   }
 
   eliminarNegocio(id_negocio: number) {
-    // TODO: Alerta => Está seguro de querer eliminar?
-    this.negocioAPIService.deleteNegocioPorID(id_negocio)
-      .then(resultadoActualizacion => {
-        // TODO: Se ha eliminado exitosamente
-
-        // Consulta los cambios en los negocios
-        return this.negocioAPIService.readNegocios();
-      })
-      .then(queryNegocios => {
-        this.negocios = queryNegocios.data as NegocioModelo[];
-      });
+    Swal.fire({
+      title: '¿Está seguro de querer eliminar el negocio?',
+      text: "Esta acción no se puede deshacer",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Eliminar',
+      cancelButtonText: 'No',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.negocioAPIService.deleteNegocioPorID(id_negocio)
+          .then(resultadoActualizacion => {
+            Swal.fire(
+              'Negocio eliminado',
+              '',
+              'success'
+            )
+            // Consulta los cambios en los negocios
+            return this.negocioAPIService.readNegocios();
+          })
+          .then(queryNegocios => {
+            this.negocios = queryNegocios.data as NegocioModelo[];
+          });
+      }
+    });
   }
 
 }
